requirements.txt
nextcord
colorama
colorlog
python-dotenv
aiosqlite